+++ Human Capital Management/HCM.Models.ViewModels/Employees/EmployeeSendEditModel.cs
@@ -27,7 +27,7 @@
         public int NationalityId { get; set; }
 
         [Required]
-        public byte DepartmentId { get; set; }
+        public int DepartmentId { get; set; }
 
         [Required]
         public int PositionId { get; set; }
